+++ hds/server/src/main/java/hds/server/ServerApplication.java
@@ -12,6 +12,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.Collections;
 import java.util.Properties;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.logging.Logger;
 
@@ -21,7 +22,7 @@ public class ServerApplication {
 	public static final int HDS_NOTARY_REPLICAS_FIRST_PORT = 9000;
 	public static final int HDS_NOTARY_REPLICAS_FIRST_CC_PORT = 10000;
 	public static final String DB_NAME_PREFIX = "hds_replica_";
-	private static final AtomicLong writeTimestamp = new AtomicLong();
+	private static final AtomicInteger writeTimestamp = new AtomicInteger();
 	private static LogManager logManager;
 
 	private static String port;
@@ -138,11 +139,11 @@ public class ServerApplication {
 		return logManager;
 	}
 
-	public static long getCurrentWriteTimestamp() {
+	public static int getCurrentWriteTimestamp() {
 		return writeTimestamp.get();
 	}
 
-	public static void setCurrentWriteTimestamp(long newWriteTimestamp) {
+	public static void setCurrentWriteTimestamp(int newWriteTimestamp) {
 		writeTimestamp.set(newWriteTimestamp);
 	}
 }
